docker/captcha-ddddocr/requirements.txt
flask==3.0.0
gunicorn==21.2.0
ddddocr==1.4.11
Pillow==10.2.0
numpy==1.26.3
